requirements.txt
bs4==0.0.1
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.31.0
